feat: Refactored middle_wares
