+++ src/data_middleware/data_middlware.rs
@@ -2,29 +2,13 @@ use async_channel::{Receiver, Sender};
 use async_trait::async_trait;
 
 #[async_trait]
-pub trait UploadSingleDataMiddleware {
+pub trait DataMiddleware {
     async fn handle_stream(&self)
         -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>>;
     async fn get_sender(
         &self,
     ) -> Sender<Result<bytes::Bytes, Box<dyn std::error::Error + Send + Sync + 'static>>>;
-    async fn get_receiver(&self) -> Receiver<bytes::Bytes>;
-}
-
-#[async_trait]
-pub trait UploadMultiDataMiddleware {
-    async fn handle_stream(&self)
-        -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>>;
-    async fn get_sender(
+    async fn get_receiver(
         &self,
-    ) -> Sender<Result<bytes::Bytes, Box<dyn std::error::Error + Send + Sync + 'static>>>;
-    async fn get_receiver(&self) -> Receiver<bytes::Bytes>;
-}
-
-#[async_trait]
-pub trait DownloadDataMiddleware {
-    async fn handle_stream(&self)
-        -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>>;
-    async fn get_sender(&self) -> Sender<bytes::Bytes>;
-    async fn get_receiver(&self) -> Receiver<bytes::Bytes>;
+    ) -> Receiver<Result<bytes::Bytes, Box<dyn std::error::Error + Send + Sync + 'static>>>;
 }
