+++ src/data_middleware/mod.rs
@@ -1,2 +1,2 @@
 pub mod data_middlware;
-pub mod empty_middleware;
+pub mod hashing_middleware;
